Move price history+repair logic into new file
